refactor(lab-x2nj): move aurora tokens to components/aurora/tokens.ts
Relocate AURORA_* constants and pillTone/controlTone helpers from
components/logs/log-theme.ts to components/aurora/tokens.ts (canonical
home per Aurora epic). Update all 15 importers across app/(admin),
components/design-system, components/logs, and components/gateway.

No re-export shim — old path is deleted immediately to prevent
two-source drift. Also drops two stale comments in settings/page.tsx
and controls-section.tsx that referenced the old log-theme module.
