+++ apps/gateway-admin/components/design-system/data-display-section.tsx
@@ -21,7 +21,7 @@ import {
   AURORA_MEDIUM_PANEL,
   AURORA_MUTED_LABEL,
   AURORA_STRONG_PANEL,
-} from '@/components/logs/log-theme'
+} from '@/components/aurora/tokens'
 import { denseRows, keyValueBlocks, metricCards } from './demo-data'
 
 export function DataDisplaySection() {
