+++ apps/gateway-admin/components/design-system/feedback-section.tsx
@@ -28,7 +28,7 @@ import {
   AURORA_MUTED_LABEL,
   AURORA_STRONG_PANEL,
   controlTone,
-} from '@/components/logs/log-theme'
+} from '@/components/aurora/tokens'
 import { feedbackModes } from './demo-data'
 
 export function FeedbackSection() {
