+++ apps/gateway-admin/components/design-system/navigation-section.tsx
@@ -34,7 +34,7 @@ import {
   AURORA_MUTED_LABEL,
   AURORA_STRONG_PANEL,
   controlTone,
-} from '@/components/logs/log-theme'
+} from '@/components/aurora/tokens'
 import { navigationTabs, sidebarPreviewItems } from './demo-data'
 
 export function NavigationSection() {
